+++ src/Hpack/Config.hs
@@ -59,8 +59,11 @@ module Hpack.Config (
 , verbatimValueToString
 , CustomSetup(..)
 , Section(..)
+, emptySectionForeignLibrary
 , Library(..)
 , Executable(..)
+, ForeignLibrary(..)
+, emptyForeignLibrary
 , Conditional(..)
 , Cond(..)
 , Flag(..)
@@ -167,6 +170,7 @@ package name version = Package {
   , packageCustomSetup = Nothing
   , packageLibrary = Nothing
   , packageInternalLibraries = mempty
+  , packageForeignLibraries = mempty
   , packageExecutables = mempty
   , packageTests = mempty
   , packageBenchmarks = mempty
@@ -176,6 +180,7 @@ package name version = Package {
 renamePackage :: String -> Package -> Package
 renamePackage name p@Package{..} = p {
     packageName = name
+  , packageForeignLibraries = fmap (renameDependencies packageName name) packageForeignLibraries
   , packageExecutables = fmap (renameDependencies packageName name) packageExecutables
   , packageTests = fmap (renameDependencies packageName name) packageTests
   , packageBenchmarks = fmap (renameDependencies packageName name) packageBenchmarks
@@ -194,6 +199,7 @@ renameDependencies old new sect@Section{..} = sect {sectionDependencies = (Depen
 packageDependencies :: Package -> [(String, DependencyInfo)]
 packageDependencies Package{..} = nub . sortBy (comparing (lexicographically . fst)) $
      (concatMap deps packageExecutables)
+  ++ (concatMap deps packageForeignLibraries)
   ++ (concatMap deps packageTests)
   ++ (concatMap deps packageBenchmarks)
   ++ maybe [] deps packageLibrary
@@ -235,6 +241,23 @@ instance Semigroup LibrarySection where
     , librarySectionGeneratedOtherModules = librarySectionGeneratedOtherModules a <> librarySectionGeneratedOtherModules b
     , librarySectionReexportedModules = librarySectionReexportedModules a <> librarySectionReexportedModules b
     , librarySectionSignatures = librarySectionSignatures a <> librarySectionSignatures b
+    }
+
+data ForeignLibrarySection = ForeignLibrarySection {
+  foreignLibrarySectionLibVersionInfo :: Last String
+, foreignLibrarySectionOtherModules :: Maybe (List Module)
+, foreignLibrarySectionGeneratedOtherModules :: Maybe (List Module)
+} deriving (Eq, Show, Generic, FromValue)
+
+instance Monoid ForeignLibrarySection where
+  mempty = ForeignLibrarySection mempty Nothing Nothing
+  mappend = (<>)
+
+instance Semigroup ForeignLibrarySection where
+  a <> b = ForeignLibrarySection {
+      foreignLibrarySectionLibVersionInfo = foreignLibrarySectionLibVersionInfo a <> foreignLibrarySectionLibVersionInfo b
+    , foreignLibrarySectionOtherModules = foreignLibrarySectionOtherModules a <> foreignLibrarySectionOtherModules b
+    , foreignLibrarySectionGeneratedOtherModules = foreignLibrarySectionGeneratedOtherModules a <> foreignLibrarySectionGeneratedOtherModules b
     }
 
 data ExecutableSection = ExecutableSection {
@@ -573,10 +596,12 @@ type SectionConfigWithDefaults asmSources cSources cxxSources jsSources a = Prod
 
 type PackageConfigWithDefaults asmSources cSources cxxSources jsSources = PackageConfig_
   (SectionConfigWithDefaults asmSources cSources cxxSources jsSources LibrarySection)
+  (SectionConfigWithDefaults asmSources cSources cxxSources jsSources ForeignLibrarySection)
   (SectionConfigWithDefaults asmSources cSources cxxSources jsSources ExecutableSection)
 
 type PackageConfig asmSources cSources cxxSources jsSources = PackageConfig_
   (WithCommonOptions asmSources cSources cxxSources jsSources LibrarySection)
+  (WithCommonOptions asmSources cSources cxxSources jsSources ForeignLibrarySection)
   (WithCommonOptions asmSources cSources cxxSources jsSources ExecutableSection)
 
 data PackageVersion = PackageVersion {unPackageVersion :: String}
@@ -587,7 +612,7 @@ instance FromValue PackageVersion where
     String s -> return (T.unpack s)
     _ -> typeMismatch "Number or String" v
 
-data PackageConfig_ library executable = PackageConfig {
+data PackageConfig_ library foreignLib executable = PackageConfig {
   packageConfigName :: Maybe String
 , packageConfigVersion :: Maybe PackageVersion
 , packageConfigSynopsis :: Maybe String
@@ -615,6 +640,8 @@ data PackageConfig_ library executable = PackageConfig {
 , packageConfigCustomSetup :: Maybe CustomSetupSection
 , packageConfigLibrary :: Maybe library
 , packageConfigInternalLibraries :: Maybe (Map String library)
+, packageConfigForeignLibraries :: Maybe (Map String foreignLib)
+, packageConfigForeignLibrary :: Maybe foreignLib
 , packageConfigExecutable :: Maybe executable
 , packageConfigExecutables :: Maybe (Map String executable)
 , packageConfigTests :: Maybe (Map String executable)
@@ -643,6 +670,8 @@ traversePackageConfig :: Traversal PackageConfig
 traversePackageConfig t p@PackageConfig{..} = do
   library <- traverse (traverseWithCommonOptions t) packageConfigLibrary
   internalLibraries <- traverseNamedConfigs t packageConfigInternalLibraries
+  foreignLibrary <- traverse (traverseWithCommonOptions t) packageConfigForeignLibrary
+  foreignLibraries <- traverseNamedConfigs t packageConfigForeignLibraries
   executable <- traverse (traverseWithCommonOptions t) packageConfigExecutable
   executables <- traverseNamedConfigs t packageConfigExecutables
   tests <- traverseNamedConfigs t packageConfigTests
@@ -650,6 +679,8 @@ traversePackageConfig t p@PackageConfig{..} = do
   return p {
       packageConfigLibrary = library
     , packageConfigInternalLibraries = internalLibraries
+    , packageConfigForeignLibrary = foreignLibrary
+    , packageConfigForeignLibraries = foreignLibraries
     , packageConfigExecutable = executable
     , packageConfigExecutables = executables
     , packageConfigTests = tests
@@ -743,6 +774,7 @@ addPathsModuleToGeneratedModules pkg
   | otherwise = pkg {
       packageLibrary = fmap mapLibrary <$> packageLibrary pkg
     , packageInternalLibraries = fmap mapLibrary <$> packageInternalLibraries pkg
+    , packageForeignLibraries = fmap mapForeignLibrary <$> packageForeignLibraries pkg
     , packageExecutables = fmap mapExecutable <$> packageExecutables pkg
     , packageTests = fmap mapExecutable <$> packageTests pkg
     , packageBenchmarks = fmap mapExecutable <$> packageBenchmarks pkg
@@ -758,6 +790,15 @@ addPathsModuleToGeneratedModules pkg
       | otherwise = lib
       where
         generatedModules = libraryGeneratedModules lib
+
+    mapForeignLibrary :: ForeignLibrary -> ForeignLibrary
+    mapForeignLibrary foreignLibrary
+      | pathsModule `elem` foreignLibraryOtherModules foreignLibrary = foreignLibrary {
+          foreignLibraryGeneratedModules = if pathsModule `elem` generatedModules then generatedModules else pathsModule : generatedModules
+        }
+      | otherwise = foreignLibrary
+      where
+        generatedModules = foreignLibraryGeneratedModules foreignLibrary
 
     mapExecutable :: Executable -> Executable
     mapExecutable executable
@@ -841,6 +882,7 @@ ensureRequiredCabalVersion inferredLicense pkg@Package{..} = pkg {
       , makeVersion [3,14] <$ guard (not (null packageExtraFiles))
       , packageLibrary >>= libraryCabalVersion
       , internalLibsCabalVersion packageInternalLibraries
+      , foreignLibsCabalVersion packageForeignLibraries
       , executablesCabalVersion packageExecutables
       , executablesCabalVersion packageTests
       , executablesCabalVersion packageBenchmarks
@@ -882,6 +924,15 @@ ensureRequiredCabalVersion inferredLicense pkg@Package{..} = pkg {
       where
         versions = libraryCabalVersion <$> Map.elems internalLibraries
 
+    foreignLibsCabalVersion :: Map String (Section ForeignLibrary) -> Maybe CabalVersion
+    foreignLibsCabalVersion = foldr max Nothing . map foreignLibCabalVersion . Map.elems
+
+    foreignLibCabalVersion :: Section ForeignLibrary -> Maybe CabalVersion
+    foreignLibCabalVersion sect = maximum [
+        makeVersion [2,0] <$ guard (foreignLibraryHasGeneratedModules sect)
+      , sectionCabalVersion (concatMap getForeignLibraryModules) sect
+      ]
+
     executablesCabalVersion :: Map String (Section Executable) -> Maybe CabalVersion
     executablesCabalVersion = foldr max Nothing . map executableCabalVersion . Map.elems
 
@@ -891,6 +942,9 @@ ensureRequiredCabalVersion inferredLicense pkg@Package{..} = pkg {
       , makeVersion [3,12] <$ guard (executableHasPackageInfoModule sect)
       , sectionCabalVersion (concatMap getExecutableModules) sect
       ]
+
+    foreignLibraryHasGeneratedModules :: Section ForeignLibrary -> Bool
+    foreignLibraryHasGeneratedModules = any (not . null . foreignLibraryGeneratedModules)
 
     executableHasGeneratedModules :: Section Executable -> Bool
     executableHasGeneratedModules = any (not . null . executableGeneratedModules)
@@ -1062,6 +1116,7 @@ data Package = Package {
 , packageCustomSetup :: Maybe CustomSetup
 , packageLibrary :: Maybe (Section Library)
 , packageInternalLibraries :: Map String (Section Library)
+, packageForeignLibraries :: Map String (Section ForeignLibrary)
 , packageExecutables :: Map String (Section Executable)
 , packageTests :: Map String (Section Executable)
 , packageBenchmarks :: Map String (Section Executable)
@@ -1081,6 +1136,24 @@ data Library = Library {
 , libraryReexportedModules :: [String]
 , librarySignatures :: [String]
 } deriving (Eq, Show)
+
+data ForeignLibrary = ForeignLibrary {
+  -- This is a hack. The 'type' field is required but we also want to exclude it
+  -- from:
+  --
+  -- if os(windows)
+  --   options: standalone
+  --
+  -- So, we treat it as if it were optional:
+  foreignLibraryType :: Maybe String
+, foreignLibraryOptions :: Maybe [String]
+, foreignLibraryLibVersionInfo :: Maybe String
+, foreignLibraryOtherModules :: [Module]
+, foreignLibraryGeneratedModules :: [Module]
+} deriving (Eq, Show)
+
+emptyForeignLibrary :: ForeignLibrary
+emptyForeignLibrary = ForeignLibrary Nothing Nothing Nothing [] []
 
 data Executable = Executable {
   executableMain :: Maybe FilePath
@@ -1125,6 +1198,42 @@ data Section a = Section {
 , sectionSystemBuildTools :: SystemBuildTools
 , sectionVerbatim :: [Verbatim]
 } deriving (Eq, Show, Functor, Foldable, Traversable)
+
+emptySectionForeignLibrary :: Section ForeignLibrary
+emptySectionForeignLibrary = Section {
+    sectionData = emptyForeignLibrary
+  , sectionSourceDirs = mempty
+  , sectionDependencies = mempty
+  , sectionPkgConfigDependencies = mempty
+  , sectionDefaultExtensions = mempty
+  , sectionOtherExtensions = mempty
+  , sectionLanguage = Nothing
+  , sectionMhsOptions = mempty
+  , sectionGhcOptions = mempty
+  , sectionGhcProfOptions = mempty
+  , sectionGhcSharedOptions = mempty
+  , sectionGhcjsOptions = mempty
+  , sectionCppOptions = mempty
+  , sectionAsmOptions = mempty
+  , sectionAsmSources = mempty
+  , sectionCcOptions = mempty
+  , sectionCSources = mempty
+  , sectionCxxOptions = mempty
+  , sectionCxxSources = mempty
+  , sectionJsSources = mempty
+  , sectionExtraLibDirs = mempty
+  , sectionExtraLibraries = mempty
+  , sectionExtraFrameworksDirs = mempty
+  , sectionFrameworks = mempty
+  , sectionIncludeDirs = mempty
+  , sectionInstallIncludes = mempty
+  , sectionLdOptions = mempty
+  , sectionBuildable = Nothing
+  , sectionConditionals = mempty
+  , sectionBuildTools = mempty
+  , sectionSystemBuildTools = mempty
+  , sectionVerbatim = mempty
+  }
 
 data Conditional a = Conditional {
   conditionalCondition :: Cond
@@ -1206,6 +1315,8 @@ expandSectionDefaults
 expandSectionDefaults formatYamlParseError userDataDir dir p@PackageConfig{..} = do
   library <- traverse (expandDefaults formatYamlParseError userDataDir dir) packageConfigLibrary
   internalLibraries <- traverse (traverse (expandDefaults formatYamlParseError userDataDir dir)) packageConfigInternalLibraries
+  foreignLibrary <- traverse (expandDefaults formatYamlParseError userDataDir dir) packageConfigForeignLibrary
+  foreignLibraries <- traverse (traverse (expandDefaults formatYamlParseError userDataDir dir)) packageConfigForeignLibraries
   executable <- traverse (expandDefaults formatYamlParseError userDataDir dir) packageConfigExecutable
   executables <- traverse (traverse (expandDefaults formatYamlParseError userDataDir dir)) packageConfigExecutables
   tests <- traverse (traverse (expandDefaults formatYamlParseError userDataDir dir)) packageConfigTests
@@ -1213,6 +1324,8 @@ expandSectionDefaults formatYamlParseError userDataDir dir p@PackageConfig{..} =
   return p{
       packageConfigLibrary = library
     , packageConfigInternalLibraries = internalLibraries
+    , packageConfigForeignLibrary = foreignLibrary
+    , packageConfigForeignLibraries = foreignLibraries
     , packageConfigExecutable = executable
     , packageConfigExecutables = executables
     , packageConfigTests = tests
@@ -1270,6 +1383,7 @@ type GlobalOptions = CommonOptions AsmSources CSources CxxSources JsSources Empt
 
 toPackage_ :: (MonadIO m, Warnings m, State m) => FilePath -> Product GlobalOptions (PackageConfig AsmSources CSources CxxSources JsSources) -> m Package
 toPackage_ dir (Product g PackageConfig{..}) = do
+  foreignLibraryMap <- toExecutableMap packageName packageConfigForeignLibraries packageConfigForeignLibrary
   executableMap <- toExecutableMap packageName packageConfigExecutables packageConfigExecutable
   let
     globalVerbatim = commonOptionsVerbatim g
@@ -1284,11 +1398,13 @@ toPackage_ dir (Product g PackageConfig{..}) = do
     toSections = maybe (return mempty) (traverse toSect)
 
     toLib = toLibrary dir packageName
+    toForeignLibraries = toSections >=> traverse (toForeignLibrary dir packageName)
     toExecutables = toSections >=> traverse (toExecutable dir packageName)
 
   mLibrary <- traverse (toSect >=> toLib) packageConfigLibrary
   internalLibraries <- toSections packageConfigInternalLibraries >>= traverse toLib
 
+  foreignLibraries <- toForeignLibraries foreignLibraryMap
   executables <- toExecutables executableMap
   tests <- toExecutables packageConfigTests
   benchmarks <- toExecutables packageConfigBenchmarks
@@ -1298,6 +1414,7 @@ toPackage_ dir (Product g PackageConfig{..}) = do
   missingSourceDirs <- liftIO $ nub . sort <$> filterM (fmap not <$> doesDirectoryExist . (dir </>)) (
        maybe [] sectionSourceDirs mLibrary
     ++ concatMap sectionSourceDirs internalLibraries
+    ++ concatMap sectionSourceDirs foreignLibraries
     ++ concatMap sectionSourceDirs executables
     ++ concatMap sectionSourceDirs tests
     ++ concatMap sectionSourceDirs benchmarks
@@ -1357,6 +1474,7 @@ toPackage_ dir (Product g PackageConfig{..}) = do
       , packageCustomSetup = mCustomSetup
       , packageLibrary = mLibrary
       , packageInternalLibraries = internalLibraries
+      , packageForeignLibraries = foreignLibraries
       , packageExecutables = executables
       , packageTests = tests
       , packageBenchmarks = benchmarks
@@ -1474,6 +1592,9 @@ getMentionedLibraryModules (LibrarySection _ _ exposedModules generatedExposedMo
 getLibraryModules :: Library -> [Module]
 getLibraryModules Library{..} = libraryExposedModules ++ libraryOtherModules
 
+getForeignLibraryModules :: ForeignLibrary -> [Module]
+getForeignLibraryModules ForeignLibrary{..} = foreignLibraryOtherModules
+
 getExecutableModules :: Executable -> [Module]
 getExecutableModules Executable{..} = executableOtherModules
 
@@ -1488,7 +1609,9 @@ removeConditionalsThatAreAlwaysFalse sect = sect {
     p = (/= CondBool False) . conditionalCondition
 
 inferModules :: (MonadIO m, State m) =>
-     FilePath
+     Bool
+     -- ^ Support only the modern handling of Paths_ modules?
+  -> FilePath
   -> String
   -> (a -> [Module])
   -> (b -> [Module])
@@ -1496,13 +1619,15 @@ inferModules :: (MonadIO m, State m) =>
   -> ([Module] -> a -> b)
   -> Section a
   -> m (Section b)
-inferModules dir packageName_ getMentionedModules getInferredModules fromData fromConditionals sect_ = do
+inferModules modernPaths dir packageName_ getMentionedModules getInferredModules fromData fromConditionals sect_ = do
   specVersion <- State.get
   let
     pathsModule :: [Module]
     pathsModule = case specVersion of
       SpecVersion v | v >= Version.makeVersion [0,36,0] -> []
-      _ -> [pathsModuleFromPackageName packageName_]
+      _
+        | modernPaths -> []
+        | otherwise -> [pathsModuleFromPackageName packageName_]
 
   removeConditionalsThatAreAlwaysFalse <$> traverseSectionAndConditionals
     (fromConfigSection fromData pathsModule)
@@ -1521,7 +1646,7 @@ inferModules dir packageName_ getMentionedModules getInferredModules fromData fr
 
 toLibrary :: (MonadIO m, State m) => FilePath -> String -> Section LibrarySection -> m (Section Library)
 toLibrary dir name =
-    inferModules dir name getMentionedLibraryModules getLibraryModules fromLibrarySectionTopLevel fromLibrarySectionInConditional
+    inferModules False dir name getMentionedLibraryModules getLibraryModules fromLibrarySectionTopLevel fromLibrarySectionInConditional
   where
     fromLibrarySectionTopLevel :: [Module] -> [Module] -> LibrarySection -> Library
     fromLibrarySectionTopLevel pathsModule inferableModules LibrarySection{..} =
@@ -1559,13 +1684,35 @@ fromLibrarySectionPlain LibrarySection{..} = Library {
   , librarySignatures = fromMaybeList librarySectionSignatures
   }
 
+getMentionedForeignLibraryModules :: ForeignLibrarySection -> [Module]
+getMentionedForeignLibraryModules (ForeignLibrarySection _ otherModules generatedModules)=
+  fromMaybeList (otherModules <> generatedModules)
+
+toForeignLibrary :: (MonadIO m, State m) => FilePath -> String -> Section ForeignLibrarySection -> m (Section ForeignLibrary)
+toForeignLibrary dir packageName_ =
+    inferModules True dir packageName_ getMentionedForeignLibraryModules getForeignLibraryModules fromForeignLibrarySection (fromForeignLibrarySection [])
+  where
+    fromForeignLibrarySection :: [Module] -> [Module] -> ForeignLibrarySection -> ForeignLibrary
+    fromForeignLibrarySection pathsModule inferableModules ForeignLibrarySection{..} =
+      (ForeignLibrary
+        -- Cabal accepts native-static or native-shared but supports only the latter:
+        (Just "native-shared")
+        Nothing
+        (getLast foreignLibrarySectionLibVersionInfo)
+        (otherModules ++ generatedModules)
+        generatedModules
+      )
+      where
+        otherModules = maybe (inferableModules ++ pathsModule) fromList foreignLibrarySectionOtherModules
+        generatedModules = maybe [] fromList foreignLibrarySectionGeneratedOtherModules
+
 getMentionedExecutableModules :: ExecutableSection -> [Module]
 getMentionedExecutableModules (ExecutableSection (Alias (Last main)) otherModules generatedModules)=
   maybe id (:) (toModule . Path.fromFilePath <$> main) $ fromMaybeList (otherModules <> generatedModules)
 
 toExecutable :: (MonadIO m, State m) => FilePath -> String -> Section ExecutableSection -> m (Section Executable)
 toExecutable dir packageName_ =
-    inferModules dir packageName_ getMentionedExecutableModules getExecutableModules fromExecutableSection (fromExecutableSection [])
+    inferModules False dir packageName_ getMentionedExecutableModules getExecutableModules fromExecutableSection (fromExecutableSection [])
   . expandMain
   where
     fromExecutableSection :: [Module] -> [Module] -> ExecutableSection -> Executable